+++ src/hooks/rooms/widgets/index.ts
@@ -1,21 +1,12 @@
-export * from './useFurnitureBackgroundColorWidget';
-export * from './useFurnitureBadgeDisplayWidget';
-export * from './useFurnitureContextMenuWidget';
-export * from './useFurnitureCraftingWidget';
-export * from './useFurnitureDimmerWidget';
-export * from './useFurnitureExchangeWidget';
-export * from './useFurnitureExternalImageWidget';
-export * from './useFurnitureFootballGateWidget';
-export * from './useFurnitureFriendFurniWidget';
-export * from './useFurnitureHighScoreWidget';
-export * from './useFurnitureInternalLinkWidget';
-export * from './useFurnitureMannequinWidget';
-export * from './useFurniturePlaylistEditorWidget';
-export * from './useFurniturePresentWidget';
-export * from './useFurnitureRentableSpaceWidget';
-export * from './useFurnitureRoomLinkWidget';
-export * from './useFurnitureSpamWallPostItWidget';
-export * from './useFurnitureStackHeightWidget';
-export * from './useFurnitureStickieWidget';
-export * from './useFurnitureTrophyWidget';
-export * from './useFurnitureYoutubeWidget';
+export * from './furniture';
+export * from './useAvatarInfoWidget';
+export * from './useChatInputWidget';
+export * from './useChatWidget';
+export * from './useDoorbellWidget';
+export * from './useFilterWordsWidget';
+export * from './useFriendRequestWidget';
+export * from './useFurniChooserWidget';
+export * from './usePetPackageWidget';
+export * from './usePollWidget';
+export * from './useUserChooserWidget';
+export * from './useWordQuizWidget';
